requirements.txt
datasets>=2.0.0
tinker
numpy
python-dotenv

